+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Upcoming
 - Documented ordering of entities in query result lists.
+- Added `hasSomeOf`, `hasAllOf`, `hasAnyOtherThan`, and `countHas` to `Entity`.
+- Implemented experimenal global component combination validators.  You can add a static `validate(entity)` method to any component type and check for valid combinations of components, throwing an error if a check fails.  *All* validation methods are executed for *all* entities whose components have changed after each system executes (not just ones that have a component of the method's host type), and the system's read entitlements are bypassed during these checks.  Entities are not validated in response to writes, so validators shouldn't look at fields.  Entities are not validated at all in the perf build.
 
 ### 0.12.1
 - Fixed world initialization and finalization to run end-of-cycle processing.